mudei numero pra int e atributos float pra bigdata(facilitar nos calculos)

+++ pedroguerra/src/main/java/com/pedroguerra/controller/ServicoController.java
@@ -13,6 +13,7 @@ import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.*;
 
+import java.math.BigDecimal;
 import java.sql.Date;
 import java.sql.SQLException;
 import java.util.List;
@@ -48,7 +49,7 @@ public class ServicoController {
     @PostMapping("/salvar-servico")
     public String salvarServico(@ModelAttribute("servico") Servico servico,
                                 @RequestParam(required = false) String operadorDroneMatricula,
-                                @RequestParam float area,
+                                @RequestParam BigDecimal area,
                                 @RequestParam("dataRelatorio") Date dataRelatorio,
                                 @RequestParam String cnpjEmpresa,
                                 @RequestParam String codigoLocalizacao,
