+++ pedroguerra/src/main/java/com/pedroguerra/model/Empresa.java
@@ -1,14 +1,16 @@
 package com.pedroguerra.model;
 
+import java.math.BigDecimal;
+
 public class Empresa {
     private String cnpj;
     private String nome;
-    private float capitalSocial;
+    private BigDecimal capitalSocial;
     private String fk_endereco_cep;
 
     public Empresa() {}
 
-    public Empresa(String cnpj, String nome, float capitalSocial, String fk_endereco_cep) {
+    public Empresa(String cnpj, String nome, BigDecimal capitalSocial, String fk_endereco_cep) {
         this.cnpj = cnpj;
         this.nome = nome;
         this.capitalSocial = capitalSocial;
@@ -21,8 +23,8 @@ public class Empresa {
     public String getNome() { return nome; }
     public void setNome(String nome) { this.nome = nome; }
 
-    public float getCapitalSocial() { return capitalSocial; }
-    public void setCapitalSocial(float capitalSocial) { this.capitalSocial = capitalSocial; }
+    public BigDecimal getCapitalSocial() { return capitalSocial; }
+    public void setCapitalSocial(BigDecimal capitalSocial) { this.capitalSocial = capitalSocial; }
 
     public String getFkEnderecoCep() { return fk_endereco_cep; }
     public void setFkEnderecoCep(String fk_endereco_cep) { this.fk_endereco_cep = fk_endereco_cep; }
